+++ terraform/ces-module/ces-preparations/variables.tf
@@ -1,0 +1,72 @@
+variable "create_namespace" {
+  description = "Flag to specify whether the namespace in variable ces_namespace shall be created"
+  type = bool
+}
+
+# component operator crd
+variable "component_operator_crd_chart" {
+  description = "The helm chart of the component crd. Optional with version like k8s/k8s-component-operator-crd:1.2.3"
+  type        = string
+}
+
+# component operator crd
+variable "blueprint_operator_crd_chart" {
+  description = "The helm chart of the blueprint crd. Optional with version like k8s/k8s-blueprint-operator-crd:1.2.3"
+  type        = string
+}
+
+# namespace of ces
+variable "ces_namespace" {
+  description = "The namespace for the CES"
+  type        = string
+}
+
+variable "helm_registry_username" {
+  description = "The username for the helm-registry"
+  type        = string
+}
+
+variable "helm_registry_password" {
+  description = "The base64-encoded password for the helm-registry"
+  type        = string
+  sensitive   = true
+}
+
+# docker credentials
+variable "docker_registry_host" {
+  description = "The host for the docker-registry"
+  type        = string
+}
+
+variable "docker_registry_username" {
+  description = "The username for the docker-registry"
+  type        = string
+}
+
+variable "docker_registry_email" {
+  description = "The email for the docker-registry"
+  type        = string
+}
+
+variable "docker_registry_password" {
+  description = "The base64-encoded password for the docker-registry"
+  type        = string
+  sensitive   = true
+}
+
+# dogu registry credentials
+variable "dogu_registry_username" {
+  description = "The username for the dogu-registry"
+  type        = string
+}
+
+variable "dogu_registry_password" {
+  description = "The base64-encoded password for the dogu-registry"
+  type        = string
+  sensitive   = true
+}
+
+variable "externalIP" {
+  description = "Contains the external IP, my overwrite the loadbalancer external ip, defaults to empty -> so the loadbalancer ip will not be patched"
+  type        = string
+}