+++ terraform/examples/ces_azure_aks/variables.tf
@@ -1,11 +1,3 @@
-variable "container_registry_secrets" {
-  description = "A list of credentials for container registries used by dogus and components. The password must be base64 encoded. The regular configuration would contain registry.cloudogu.com as url."
-  type        = list(object({
-    url      = string
-    username = string
-    password = string
-  }))
-}
 
 variable "dogu_registry_username" {
   description = "The username for the dogu-registry"
@@ -18,16 +10,26 @@ variable "dogu_registry_password" {
   sensitive   = true
 }
 
-variable "dogu_registry_endpoint" {
-  description = "The endpoint for the dogu-registry"
+# docker credentials
+variable "docker_registry_host" {
+  description = "The host for the docker-registry"
   type        = string
-  default     = "https://dogu.cloudogu.com/api/v2/dogus"
 }
 
-variable "dogu_registry_url_schema" {
-  description = "The URL schema for the dogu-registry ('default' or 'index')"
+variable "docker_registry_username" {
+  description = "The username for the docker-registry"
   type        = string
-  default     = "default"
+}
+
+variable "docker_registry_email" {
+  description = "The email for the docker-registry"
+  type        = string
+}
+
+variable "docker_registry_password" {
+  description = "The base64-encoded password for the docker-registry"
+  type        = string
+  sensitive   = true
 }
 
 variable "helm_registry_host" {
@@ -40,18 +42,6 @@ variable "helm_registry_schema" {
   description = "The schema for the helm-registry"
   type        = string
   default     = "oci"
-}
-
-variable "helm_registry_plain_http" {
-  description = "A flag which indicates if the component-operator should use plain http for the helm-registry"
-  type        = bool
-  default     = false
-}
-
-variable "helm_registry_insecure_tls" {
-  description = "A flag which indicates if the component-operator should use insecure TLS for the helm-registry"
-  type        = bool
-  default     = false
 }
 
 variable "helm_registry_username" {
@@ -101,8 +91,6 @@ variable "dogus" {
   default     = [
     "official/ldap",
     "official/postfix",
-    "k8s/nginx-static",
-    "k8s/nginx-ingress",
     "official/cas",
     "official/jenkins",
     "official/nexus",
@@ -110,50 +98,95 @@ variable "dogus" {
   ]
 }
 
+# component operator crd
 variable "component_operator_crd_chart" {
-  description = "The helm chart of the component crd. Optional with version like k8s/k8s-component-operator-crd:1.2.1"
-  type = string
-  default = "k8s/k8s-component-operator-crd:latest"
+  description = "The helm chart of the component crd. Optional with version like k8s/k8s-component-operator-crd:1.2.3"
+  type        = string
+  default     = "k8s/k8s-component-operator-crd:1.10.1"
 }
 
-variable "component_operator_chart" {
-  description = "The helm chart of the component operator. Optional with version like k8s/k8s-component-operator:1.2.1"
-  type = string
-  default = "k8s/k8s-component-operator:latest"
+# component operator crd
+variable "component_operator_image" {
+  description = "The Image:Version of the component operator. Optional with version like cloudogu/k8s-component-operator:1.10.0"
+  type        = string
+  default     = "cloudogu/k8s-component-operator:1.10.1"
 }
 
+# List of cómponents, backup components and monitoring components
 variable "components" {
-  description = "A list of components to install, optional with version like k8s/k8s-dogu-operator:3.0.1"
-  type = list(string)
-  default = [
-    "k8s/k8s-dogu-operator",
-    "k8s/k8s-dogu-operator-crd",
-    "k8s/k8s-service-discovery",
-  ]
+  description = "A list of credentials for container registries used by dogus and components. The password must be base64 encoded. The regular configuration would contain registry.cloudogu.com as url."
+  type = object ({
+    components = list(object({
+      namespace = string
+      name = string
+      version = string
+      helmNamespace = optional(string)
+      disabled = optional(bool, false)
+      valuesObject = optional(any, null)
+    }))
+    backup = object ({
+      enabled = bool
+      components = list(object({
+        namespace = string
+        name = string
+        version = string
+        helmNamespace = optional(string)
+        disabled = optional(bool, false)
+        valuesObject = optional(any, null)
+      }))
+    })
+    monitoring = object ({
+      enabled = bool
+      components = list(object({
+        namespace = string
+        name = string
+        version = string
+        helmNamespace = optional(string)
+        disabled = optional(bool, false)
+        valuesObject = optional(any, null)
+      }))
+    })
+  })
+  default = {
+    components = [
+      { namespace = "ecosystem", name = "k8s-dogu-operator-crd", version = "2.9.0" },
+      { namespace = "ecosystem", name = "k8s-dogu-operator", version = "3.13.0" },
+      { namespace = "ecosystem", name = "k8s-service-discovery", version = "3.0.0" },
+      { namespace = "ecosystem", name = "k8s-blueprint-operator-crd", version = "3.0.1", disabled = true},
+      { namespace = "ecosystem", name = "k8s-blueprint-operator", version = "2.7.0" },
+      { namespace = "ecosystem", name = "k8s-ces-gateway", version = "1.0.1" },
+      { namespace = "ecosystem", name = "k8s-ces-assets", version = "1.0.1" },
+      { namespace = "ecosystem", name = "k8s-ces-control", version = "1.7.1", disabled = true },
+      { namespace = "ecosystem", name = "k8s-debug-mode-operator-crd", version = "0.2.3"},
+      { namespace = "ecosystem", name = "k8s-debug-mode-operator", version = "0.3.0"},
+      { namespace = "ecosystem", name = "k8s-support-mode-operator-crd", version = "0.2.0", disabled = true },
+      { namespace = "ecosystem", name = "k8s-support-mode-operator", version = "0.3.0", disabled = true },
+    ]
+    backup = {
+      enabled = true
+      components = [
+        { namespace = "ecosystem", name = "k8s-backup-operator-crd", version = "1.6.0" },
+        { namespace = "ecosystem", name = "k8s-backup-operator", version = "1.6.0" },
+        { namespace = "ecosystem", name = "k8s-velero", version = "10.0.1-5" },
+      ]
+    }
+    monitoring = {
+      enabled = true
+      components = [
+        { namespace = "ecosystem", name = "k8s-prometheus", version = "75.3.5-3" },
+        { namespace = "ecosystem", name = "k8s-minio", version = "2025.6.13-2" },
+        { namespace = "ecosystem", name = "k8s-loki", version = "3.3.2-4" },
+        { namespace = "ecosystem", name = "k8s-promtail", version = "2.9.1-9" },
+        { namespace = "ecosystem", name = "k8s-alloy", version = "1.1.2-1" },
+      ]
+    }
+  }
 }
 
 variable "ces_fqdn" {
   description = "Fully qualified domain name of the EcoSystem, e.g. 'www.ecosystem.my-domain.com'"
   type        = string
   default     = ""
-}
-
-variable "setup_chart_namespace" {
-  description = "The namespace of k8s-ces-setup chart"
-  type        = string
-  default     = "k8s"
-}
-
-variable "setup_chart_version" {
-  description = "The version of the k8s-ces-setup chart"
-  type        = string
-  default     = "4.1.1"
-}
-
-variable "resource_patches_file" {
-  description = "The location of a file containing resource-patches for the CES installation. The file-path is relative to the root-module-location"
-  type        = string
-  default     = "resource_patches.yaml"
 }
 
 variable "node_pool_name" {